+++ LePong/utils.h
@@ -1,7 +1,7 @@
 #include <raylib.h>
 
 float getDistanceFromMiddle(Rectangle rectangle, Vector2 point);
-float getHalf(int size);
+float getHalf(float size);
 float getWithFrameTime(float value);
 bool checkRectangleCollision(Rectangle rectangle1, Rectangle rectangle2);
 float clamp(float value, float min, float max);
